fake-news-detection-webapp/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
torch>=2.0.0
transformers>=4.27.4
python-multipart>=0.0.6
newspaper3k>=0.2.8
beautifulsoup4>=4.11.2
jinja2>=3.1.2
pytest>=7.3.1
httpx>=0.24.0
aiofiles>=23.1.0
scikit-learn>=1.2.2